+++ zipkin/src/main/java/com/linecorp/armeria/internal/tracing/PingPongExtra.java
@@ -4,11 +4,9 @@ import brave.propagation.TraceContext;
 
 /** Hack to allow us to peek inside a current trace context implementation */
 public final class PingPongExtra {
-    private boolean pong;
-
     /** If the input includes only this extra, set {@link #isPong() pong = true} */
     public static boolean maybeSetPong(TraceContext context) {
-        if (context.extra().size() != 1) {
+        if (context.extra().size() == 1) {
             Object extra = context.extra().get(0);
             if (extra instanceof PingPongExtra) {
                 ((PingPongExtra) extra).pong = true;
@@ -18,7 +16,9 @@ public final class PingPongExtra {
         return false;
     }
 
-    public boolean isPong() {
+    private boolean pong;
+
+    boolean isPong() {
         return pong;
     }
 }
