+++ zipkin/src/main/java/com/linecorp/armeria/internal/tracing/SpanContextUtil.java
@@ -28,19 +28,25 @@ import brave.propagation.TraceContext;
 
 public final class SpanContextUtil {
 
+    /**
+     * Trace context scopes are often wrapped, for example to decorate MDC. This creates a dummy scope to
+     * ensure {@link RequestContextCurrentTraceContext#INSTANCE} is configured, even if it is wrapped.
+     */
     public static void ensureScopeUsesRequestContext(Tracing tracing) {
-        PingPongExtra extra = new PingPongExtra();
+        final PingPongExtra extra = new PingPongExtra();
         // trace contexts are not recorded until Tracer.toSpan, so this won't end up as junk data
-        TraceContext dummyContext = TraceContext.newBuilder().traceId(1).spanId(1)
-                                                .extra(Collections.singletonList(extra)).build();
-        boolean scopeUsesRequestContext;
+        final TraceContext dummyContext = TraceContext.newBuilder().traceId(1).spanId(1)
+                                                      .extra(Collections.singletonList(extra)).build();
+        final boolean scopeUsesRequestContext;
         try (CurrentTraceContext.Scope scope = tracing.currentTraceContext().newScope(dummyContext)) {
             scopeUsesRequestContext = extra.isPong();
         }
         if (!scopeUsesRequestContext) {
-            throw new IllegalStateException("Please initialize Tracing.Builder.currentTraceContext("
-                                            + "new " + RequestContextCurrentTraceContext.class.getSimpleName()
-                                            + "()");
+            throw new IllegalStateException(
+                    "Tracing.currentTraceContext is not a " + RequestContextCurrentTraceContext.class
+                            .getSimpleName() + " scope. " +
+                    "Please call Tracing.Builder.currentTraceContext" + RequestContextCurrentTraceContext.class
+                            .getSimpleName() + ".INSTANCE).");
         }
     }
